# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kasten",
    "version": "0.31.0"
  },
  "paths": {
    "/agent/notes/{path}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Read Note",
        "description": "Read one note, with the digest a conditional write presents back.",
        "operationId": "read_note_agent_notes__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agent"
        ],
        "summary": "Save Note",
        "description": "Write one note, creating it when `sha` is absent and the note is.",
        "operationId": "save_note_agent_notes__path__put",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Save"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/notes": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Notes",
        "description": "List every note in the vault, or every note in one folder of it.",
        "operationId": "list_notes_agent_notes_get",
        "parameters": [
          {
            "name": "folder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Folder"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Notes Agent Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/search": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Search Notes",
        "description": "Find every line in the vault holding `q`, ignoring case.",
        "operationId": "search_notes_agent_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "archive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hit"
                  },
                  "title": "Response Search Notes Agent Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/graph": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Graph",
        "description": "The notes as a graph, narrowed by a filter or asked a pattern.",
        "operationId": "graph_agent_graph_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "around",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Around"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 0,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "archive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Graph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent/notes/{path}/append": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Append Note",
        "description": "Add a line to the end of one note, creating it when there is none.",
        "operationId": "append_note_agent_notes__path__append_post",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Append"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/dump": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Dump",
        "description": "Add a paragraph to the `## Dump` of the daily note for `date`, making either.",
        "operationId": "dump_agent_dump_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dump"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/files/{path}/fetch": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Fetch File",
        "description": "Put the file at a public `url` into the vault at `path`, never over one there.",
        "operationId": "fetch_file_agent_files__path__fetch_post",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Fetch"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Filed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/files/{path}": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Upload File",
        "description": "Put the raw body into the vault at `path` as a book or an image, never over one there.\n\nRaw rather than multipart, as `POST /api/assets/{path}` takes it, so `curl\n--data-binary @paper.pdf` is the whole client. `openapi_extra` because the\nbody is read off the request, which FastAPI cannot see to describe.",
        "operationId": "upload_file_agent_files__path__post",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Filed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "summary": "Search Files",
        "description": "Find every line in the vault containing `q`, ignoring case.\n\nA literal match, not a regex and not a fuzzy one. The client ranks what\ncomes back, which is what makes the finder feel fuzzy without asking a\nsubsequence match to mean something over prose, where it matches everything.\n\n`archive` walks the archive folder too. Off by default, because what is\nfiled away is not usually what is being looked for, and the cap on what\ncomes back means an archive left in would eventually push live notes out of\nthe answer rather than merely lengthening it.",
        "operationId": "search_files_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "archive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Search Files Api Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/graph": {
      "get": {
        "summary": "Show Graph",
        "description": "The notes as a graph, narrowed by a filter or asked a pattern.\n\nThe query is text and the backend reads it, so the pane, this route and the\nagent's tool share one parser. `around` names a note and draws only what is\nwithin `depth` links of it, which is the local graph. A query that cannot\nbe read is a 400 whose detail says why, in words the pane shows as they are.\n\n`archive` walks the archive folder too, off for the reason it is off on a\nsearch: a finished project's links are true of the note and not of the\nvault as it is now.",
        "operationId": "show_graph_api_graph_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "around",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Around"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 0,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "archive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Graph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/assets/{path}": {
      "get": {
        "summary": "Read Asset",
        "description": "Read one book or image out of the vault.\n\nThe only endpoint that answers with bytes rather than with a note. It\nresolves a path, checks a suffix and streams a file; it never opens what it\nsends, so nothing here knows what an epub or a png is beyond its name.\n\nNo `media_type`: `mimetypes` answers `application/epub+zip` for `.epub` and\n`image/png` for `.png`, and starlette reads it off the path. `Range` comes\nfree with `FileResponse` and nothing uses it, the client asking for the whole\nfile once.\n\nThe `POST` below is the other half. Between them a book or an image gets into\nthe vault and back out of it without a terminal.",
        "operationId": "read_asset_api_assets__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Write Asset",
        "description": "Put one book or image into the vault, and never over one already there.\n\nBoth decorator arguments earn their place. Without `status_code` the runtime\nanswers 201 while OpenAPI documents a 200; without `response_class` FastAPI\ndocuments the 201 as JSON carrying an empty schema, and\n`openapi-typescript` turns that into a body for a response that has none.",
        "operationId": "write_asset_api_assets__path__post",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Asset",
        "description": "Take one image out of the vault, and hold it in the trash.\n\nThe note delete's route read for an image, down to the entry it answers\nwith: `move_to_trash` moves whatever path it is handed, and the trash names\nan entry after where it came from, so an image sits in there beside the notes\nand `PATCH /api/trash/{entry}` puts it back with no rule of its own. That is\nwhat makes a mistyped key cost a keypress rather than a picture in a vault\nwith no other delete.\n\nImages alone, though `resolve_asset` answers for books too. A book travels\nwith the note beside it, and which of the pair a delete should take is a\ndecision nothing here has made; the file tree, which is what this route\nserves, lists images and no books. So a book is a 404 like any other path\nthis does not serve, whichever of `BOOK_SUFFIXES` it ends in.\n\nThe notes pointing at the image are left alone, for the reason the note\ndelete leaves `[[link]]`s alone: rewriting them to say the picture is gone\nwould be the one edit a restore could not take back. The editor draws a\nreference to a file that is not there as a picture that will not load.",
        "operationId": "delete_asset_api_assets__path__delete",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrashEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dump": {
      "post": {
        "summary": "Capture Dump",
        "description": "Add one paragraph to the `## Dump` section of a day's note.\n\nThe daily ritual reads that section at 17:00 and plans the next day from it,\nand this is the way into it from anywhere else: a key in the app, a page on\na phone, a share from another app. The note and the section are made where\nthe vault has neither, the note exactly as `<leader>gd` would make it.\n\nA route of its own rather than a read and a `PUT` from the client, because\nthe read and the write happen under one hold of the lock. A phone and an\neditor autosaving the same note would otherwise race, and the loser's\nparagraph would vanish.\n\n`agent.dump` is the whole of it, and `POST /agent/dump` calls the same\nfunction: one rule about where a thought lands, whoever sent it. Text with\nnothing in it after the trim is a 422, answered by the handler registered\nabove.",
        "operationId": "capture_dump_api_dump_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dump"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Fetch": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "Fetch",
        "description": "Where to read a file from, for a caller holding a link rather than the bytes."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hit": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "line": {
            "type": "integer",
            "title": "Line"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "path",
          "line",
          "text"
        ],
        "title": "Hit",
        "description": "One matching line: where it lives, which line it is, and what it says.\n\nIts own model rather than `main.SearchHit`, because nothing here may import\nfrom `main.py`: that would point the dependency at the surface instead of\naway from it, and the MCP tools import this module too."
      },
...
      "NoteRead": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "sha": {
            "type": "string",
            "title": "Sha"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content",
          "sha"
        ],
        "title": "NoteRead",
        "description": "One note, and the digest a conditional write presents back."
      },
...
      "Save": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "Save",
        "description": "A whole note, and the digest of the one the caller read."
      },
...
    }
  }
}